+++ experiments/one_stimulus/cluster/launch_job.pbs
@@ -1,21 +1,22 @@
 #!/bin/bash
 
 
-#PBS -l walltime=00:20:00
+#PBS -l walltime=00:30:00
 
 #PBS -l nodes=1:ppn=1
 #PBS -l pmem=1gb
 #PBS -u fbenurea
 #PBS -q inria
 
-#PBS -o /scratch/fbenurea/monkeys/sgstim/data/data_sgstim.out
-#PBS -e /scratch/fbenurea/monkeys/sgstim/data/data_sgstim.err
 
 cd $PBS_O_WORKDIR
+
+#PBS -o ~/logs/logout.$PBS_ARRAYID
+#PBS -e ~/logs/logerr.$PBS_ARRAYID
 
 # replace "monkeys" by your pyenv's virtual environment name.
 # if you do not have pyenv, python version will be unchanged.
 
 export PYENV_VERSION=monkeys
-python run.py ${PBS_ARRAYID}
+python run.py $1 ${PBS_ARRAYID}
 unset PYENV_VERSION
