+++ experiments/one_stimulus/cluster/prepare.py
@@ -41,7 +41,7 @@ def write_changes(idx, changes):
     with open(os.path.join(path, filename), 'w') as fp:
         json.dump(changes, fp, indent=4, ensure_ascii=False)
 
-def copy_json(path):
+def copy_json(path='.'):
     """Copy the base model and task json file"""
     # model
     model_path_from = os.path.join(path, config.params['model_base'])
@@ -57,35 +57,18 @@ def copy_json(path):
         os.makedirs(os.path.dirname(task_path_to))
     shutil.copyfile(task_path_from, task_path_to)
 
-def param_generator(param_ranges):
+def param_generator(param_ranges=None):
     """Generate all the parameter ranges.
 
     :param params_ranges: a dict of values that the parameters can take,
                           e.g. {'rl': [0.01, 0.02, 0.03], 'hebbian': [True, False]}
     """
+    if param_ranges is None:
+        param_ranges = config.params['params']
+
     sorted_keys = sorted(param_ranges.keys())
     combinations = itertools.product(*[param_ranges[k] for k in sorted_keys])
     param_comb = []
     for values in combinations:
         param_comb.append({k: v for k, v in zip(sorted_keys, values)})
     return param_comb
-
-def compute_changes(params):
-    changes = {'task': {'single': {}, 'choice': {}}}
-    changes['task']['single']['n_trials'] = params['n_trials']
-
-    cue_freq_a = params['cue_freq_a']
-    changes['task']['single']['cue'] = [cue_freq_a, 1.0 - cue_freq_a, 0, 0]
-
-    rwd_freq_a = params['rwd_freq_a']
-    for name in ['single', 'choice']:
-        changes['task'][name]['rwd'] = [rwd_freq_a, 1.0 - rwd_freq_a, 0.0, 0.0]
-
-    return changes
-
-
-if __name__ == "__main__":
-    copy_json('.')
-    for idx, params in enumerate(param_generator(config.params['params'])):
-        changes = compute_changes(params)
-        write_changes(idx, changes)
