+++ experiments/one_stimulus/cluster/run_all.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 # qsub needs Python 2 on the avakas cluster
 export PYENV_VERSION=2.7.12
-INDICES="0-2199"
+INDICES="0-149"
 
-if (( "$#" > 0 ))
+if (( "$#" > 1 ))
 then
-  INDICES=$1;
+  INDICES=$2;
 fi
 
-CMD="qsub -t ${INDICES} launch_job.pbs"
+CMD="qsub -t ${INDICES} -F \"$1\" launch_job.pbs"
 echo $CMD
 eval $CMD
 unset PYENV_VERSION
